Change: Move download context manager to pontos.helper
It is a very generic function to download content/data from a http url
to a local path.
