+++ pontos/github/api.py
@@ -16,19 +16,15 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from collections import defaultdict
-from contextlib import contextmanager
 from enum import Enum
 from pathlib import Path
 from typing import Callable, Dict, Generator, Iterable, Iterator, List, Optional
 
 import httpx
 
-from pontos.helper import DownloadProgressIterable
+from pontos.helper import DownloadProgressIterable, download
 
 DEFAULT_GITHUB_API_URL = "https://api.github.com"
-
-DEFAULT_TIMEOUT = 1000
-DEFAULT_CHUNK_SIZE = 4096
 
 
 class FileStatus(Enum):
@@ -36,47 +32,6 @@ class FileStatus(Enum):
     DELETED = "deleted"
     MODIFIED = "modified"
     RENAMED = "renamed"
-
-
-@contextmanager
-def download(
-    url: str,
-    destination: Optional[Path] = None,
-    *,
-    chunk_size: int = DEFAULT_CHUNK_SIZE,
-    timeout: int = DEFAULT_TIMEOUT,
-) -> Generator[DownloadProgressIterable, None, None]:
-    """Download file in url to filename
-
-    Arguments:
-        url: The url of the file we want to download
-        destination: Path of the file to store the download in. If set it will
-                     be derived from the passed URL.
-        chunk_size: Download file in chunks of this size
-        timeout: Connection timeout
-
-    Raises:
-        HTTPError if the request was invalid
-
-    Returns:
-        A DownloadProgressIterator that yields the progress of the download in
-        percent for each downloaded chunk or None for each chunk if the progress
-        is unknown.
-    """
-    destination = Path(url.split("/")[-1]) if not destination else destination
-
-    with httpx.stream(
-        "GET", url, timeout=timeout, follow_redirects=True
-    ) as response:
-        response.raise_for_status()
-
-        total_length = response.headers.get("content-length")
-
-        yield DownloadProgressIterable(
-            response.iter_bytes(chunk_size=chunk_size),
-            destination,
-            total_length,
-        )
 
 
 def _get_next_url(response) -> Optional[str]:
